Implemented logout mechanism and added a navigation bar for enhanced user navigation experience
