+++ src/main/java/com/github/adhambadawi/minisurveymonkey/controller/SurveyController.java
@@ -5,8 +5,6 @@ import com.github.adhambadawi.minisurveymonkey.model.User;
 import com.github.adhambadawi.minisurveymonkey.service.SurveyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,7 +12,6 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/api/surveys")
 public class SurveyController {
-
 
     private final SurveyService surveyService;
 
@@ -79,16 +76,5 @@ public class SurveyController {
     @PutMapping("/close/{id}")
     public void closeSurvey(@PathVariable Long id) {
         surveyService.closeSurvey(id);
-    }
-
-    @GetMapping("/survey/{id}/participate")
-    public String participateSurvey(@PathVariable Long id, Model model) {
-        Optional<Survey> optionalSurvey = surveyService.getSurveyById(id);
-        if (optionalSurvey.isPresent()) {
-            model.addAttribute("surveyId", id);
-            return "fillSurvey";
-        } else {
-            return "redirect:/";
-        }
     }
 }
