+++ src/main/resources/templates/header.html
@@ -1,0 +1,39 @@
+<!-- header.html -->
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="https://www.thymeleaf.org/extras/spring-security">
+<head>
+    <!-- Head content -->
+</head>
+<body>
+<header th:fragment="header">
+    <!-- Navigation Links -->
+    <nav>
+        <a th:href="@{/}">Home</a> |
+        <span sec:authorize="isAuthenticated()">
+                Welcome, <span sec:authentication="name"></span> |
+                <form th:action="@{/logout}" method="post" style="display:inline;">
+                    <!-- CSRF Token -->
+                    <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
+                    <button type="submit" style="
+                        background:none;
+                        border:none;
+                        padding:0;
+                        margin:0;
+                        font-size:1em;
+                        color:blue;
+                        text-decoration:underline;
+                        cursor:pointer;
+                    ">
+                        Logout
+                    </button>
+                </form>
+            </span>
+        <span sec:authorize="isAnonymous()">
+                <a th:href="@{/login}">Login</a> |
+                <a th:href="@{/register}">Register</a>
+            </span>
+    </nav>
+</header>
+</body>
+</html>
