+++ src/main/resources/templates/login.html
@@ -5,6 +5,7 @@
     <title>Login</title>
 </head>
 <body>
+<div th:replace="~{header :: header}"></div>
 <h1>Login</h1>
 <form th:action="@{/login}" method="post">
     <!-- CSRF Token -->
@@ -20,7 +21,7 @@
     Invalid username or password.
 </p>
 <p th:if="${param.logout}">
-    You have been logged out.
+    You have been logged out successfully.
 </p>
 <p>
     Don't have an account? <a th:href="@{/register}">Register here</a>.
