+++ src/main/resources/templates/surveyForm.html
@@ -8,6 +8,8 @@
 </head>
 <body>
 
+<div th:replace="~{header :: header}"></div>
+
 <h1>Create Survey</h1>
